Generate material classes and fix code.
The classes are generated by the scheme file located in the server-project. This will help to develop multiple applications and libraries. The generated classes should not be changed, because the changed will be gone after re-generating the classes.
The generator and a README is also located in the server directory. Just have a look at https://github.com/go-messaging-service/goms-server/tree/master/scheme